SteamVR: Fixed the aspect ratio and size of the mirror window image.
Lots of math and trial-and-error, but I think this is the right way to display
the image. Also had to acquire the size of the actual window used to display
the image since that can be different from both the in-game resolution *and*
the backbuffer.

+++ impl11/shaders/PixelShaderTextureCommon.h
@@ -10,7 +10,7 @@ cbuffer ConstantBuffer : register(b0)
 	uint bIsLaser;				// 1 for Laser objects, setting this to 2 will make them brighter (intended for 32-bit mode)
 	uint bIsLightTexture;		// 1 if this is a light texture, 2 will make it brighter (intended for 32-bit mode)
 	uint bIsEngineGlow;			// 1 if this is an engine glow textures, 2 will make it brighter (intended for 32-bit mode)
-	uint ps_unused1;				// Formerly this was bIsSun
+	uint ps_unused1;
 	// 32 bytes
 
 	float fBloomStrength;		// General multiplier for the bloom effect
@@ -25,7 +25,7 @@ cbuffer ConstantBuffer : register(b0)
 
 	float fSpecVal, fDisableDiffuse;
 	uint special_control;
-	float ps_unused2;
+	uint ps_unused2;
 	// 80 bytes
 };
